Allows to navigate the suggestions as if they were in a loop
We want to provide intuitive keyboard navigation inside the navigation. To support this the user can press the up arrow when no suggestion is highlighted to directly jump to the bottommost suggestion. Also when the user has highlighted the bottommost suggestion he can press the down arrow to jump to no highlighted suggestion so the next down press highlights the topmost suggestion.

Fixes: OX-6121